Complete overhaul of backend. Implemented spring security and swagger ui for development and security purposes as well.

+++ src/main/java/com/saladbar/rinx/repository/GoalieRepository.java
@@ -3,5 +3,5 @@ package com.saladbar.rinx.repository;
 import com.saladbar.rinx.entity.Goalie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoalieRepository extends JpaRepository<Goalie, Integer> {
+public interface GoalieRepository extends JpaRepository<Goalie, Long> {
 }
