+++ src/main/java/com/saladbar/rinx/repository/SkaterRepository.java
@@ -3,5 +3,5 @@ package com.saladbar.rinx.repository;
 import com.saladbar.rinx.entity.Skater;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkaterRepository extends JpaRepository<Skater, Integer> {
+public interface SkaterRepository extends JpaRepository<Skater, Long> {
 }
